Better update latency handling and cli ping
Instead of sending a latency update on each ping, they are gathered
together and sent in one request. This reduces the number of HTTP
requests on Kahu but also frees up network resources so that KeKahu is
not blocking Alia operations.

Created a SendNPings function on the client that is meant to be used
from the command line only. It looks up the neighbors only once, then
sends the n pings, updating the metrics and printing to the CLI, but not
updating Kahu.

Fixes #1
Fixes #2